Add catalog class subcommand (#21)
* Add detail and summary table formatters

These are specialized tablewriters that are optimized for displaying
either a list of results, or a single item. A lot of the "noise" from
the borders have been removed.

* Move catalog list to catalog class list

We are no longer going to put verbs directly under catalog. Instead we
are using the "catalog" command to group together catalog level concepts,
class and plan, vs cluster level concepts, like instance and binding.

* Implement catalog class get [--traverse]