Split cli from library(app)
I am calling the library for svcat an app, against because I'm running
out of unique names for "client".

The app includes our sdk, and will add new functionality that probably
won't be welcome in the sdk, such as new commands to check if
the cluster has service catalog properly configured, installing the
service catalog on a new cluster, etc.